% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{actuals}
\alias{actuals}
\title{Actual Draftkings Fantasy Points for the nhl_players set}
\format{
A data table with 308 rows and 8 variables:
\describe{
  \item{fullname}{Player's full name}
  \item{id}{associated DK ID}
  \item{actuals}{Player's actual fantasy points}
}
}
\usage{
actuals
}
\description{
A dataset containing the names, ids, and true fantasy points for
the same draftkings slate used in nhl_players.
}
\keyword{datasets}
