% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{add_additional_constraints,optimizer-method}
\alias{add_additional_constraints,optimizer-method}
\title{Add additional constraints}
\usage{
\S4method{add_additional_constraints}{optimizer}(object)
}
\arguments{
\item{object}{Optimizer model object}
}
\value{
The Optimizer object with an updated model.
}
\description{
Add additional constraints
}
\details{
Adds constraints from object configuration.
}
