% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{add_block_constraint}
\alias{add_block_constraint}
\title{Block Players Constraint}
\usage{
add_block_constraint(model, players)
}
\arguments{
\item{model}{Model object}

\item{players}{List of player objects}
}
\description{
Block Players Constraint
}
\keyword{internal}
