% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{add_lock_constraint}
\alias{add_lock_constraint}
\title{Lock Players Constraint}
\usage{
add_lock_constraint(model, lock_vector)
}
\arguments{
\item{model}{Model object}

\item{lock_vector}{Vector where 1 indicates 'block'}
}
\description{
Lock Players Constraint
}
\keyword{internal}
