% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-model.R
\name{add_max_overlap_constraint}
\alias{add_max_overlap_constraint}
\title{Max Overlap constraint}
\usage{
add_max_overlap_constraint(model, roster_indx_list, max_overlap)
}
\arguments{
\item{model}{The model to further constrain}

\item{roster_indx_list}{the index of players to constrain}

\item{max_overlap}{the Maximum player overlap across lineups}
}
\description{
Max Overlap constraint
}
\keyword{internal}
