% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-model.R
\name{add_max_share_constraint}
\alias{add_max_share_constraint}
\title{Max Share Across Lineups}
\usage{
add_max_share_constraint(model, roster_indx, max_share)
}
\arguments{
\item{model}{The model to further constrain}

\item{roster_indx}{the index of players to constrain}

\item{max_share}{Number indicating how many players are allowed to be shared across lineups}
}
\description{
Max Share Across Lineups
}
\keyword{internal}
