% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{add_player,optimizer-method}
\alias{add_player,optimizer-method}
\alias{add_player}
\title{Method for adding a player to Optimizer object}
\usage{
\S4method{add_player}{optimizer}(object, player)
}
\arguments{
\item{object}{an S4 optimizer object}

\item{player}{an object of class Player}
}
\value{
Updated optimizer object
}
\description{
Method for adding a player to Optimizer object
}
