% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{add_players_from_csv,optimizer-method}
\alias{add_players_from_csv,optimizer-method}
\title{Add players to optimizer from CSV}
\usage{
\S4method{add_players_from_csv}{optimizer}(object, filepath)
}
\arguments{
\item{object}{An optimizer model object}

\item{filepath}{Filepath location of the CSV}
}
\value{
Optimizer model object with slot \code{players} filled
}
\description{
Add players to optimizer from CSV
}
\examples{
\dontrun{
opt <- optimizer(site = 'DRAFTKINGs', sport = 'HOCKEY', contest_type = 'CLASSIC')
opt <- add_players_from_csv(object = opt, filepath = '/Path/to/file.csv')
}

}
