% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{add_players_from_df,optimizer-method}
\alias{add_players_from_df,optimizer-method}
\title{Add players to optimizer from a data.frame}
\usage{
\S4method{add_players_from_df}{optimizer}(object, df)
}
\arguments{
\item{object}{An optimizer model object}

\item{df}{a data.frame of players to add to the model}
}
\value{
Optimizer model object with slot \code{players} filled
}
\description{
Add players to optimizer from a data.frame
}
\examples{
\dontrun{
opt <- create_optimizer(site = 'DRAFTKINGS', sport = 'HOCKEY', contest_type = 'CLASSIC')
dat <- data.frame(first_name = c('Sidney','Alex'),
                 last_name = c('Crosby','Ovechkin'),
                 fpts = c(18,20),
                 team = c('PIT','WAS'),
                 position = c('C','W'),
                 salary = c(8000,9000),
                 game_info = c('this will be','ignored anyway'),
                 max_exposure = c(0.3, 0.6))
opt <- add_players_from_df(object = opt, df = dat)
}

}
