% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-model.R
\name{add_position_constraint}
\alias{add_position_constraint}
\title{Position constraint}
\usage{
add_position_constraint(model, position_vector, roster_key, flex_positions)
}
\arguments{
\item{model}{OMPR Model object}

\item{position_vector}{Vector of positions to consider}

\item{roster_key}{List containing}

\item{flex_positions}{Positions to be considered for flex}
}
\description{
Position constraint
}
\keyword{internal}
