% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{add_team_stack,optimizer-method}
\alias{add_team_stack,optimizer-method}
\title{Add a Team Stack}
\usage{
\S4method{add_team_stack}{optimizer}(object, positions)
}
\arguments{
\item{object}{An optimizer model object}

\item{positions}{Positions to stack within same team}
}
\value{
Updated optimizer object
}
\description{
Add a Team Stack
}
\examples{
\dontrun{
opt <- create_optimizer(site = 'DRAFTKINGs', sport = 'HOCKEY', contest_type = 'CLASSIC')
opt <- add_players_from_csv(object = opt, filepath = '/Path/to/file.csv')

# Add team stack, requiring a Center and Two Wingers from the same team
opt <- add_team_stack(object = opt, positions = c('C','W','W'))
}

}
