% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-model.R
\name{add_unique_lineup_constraint}
\alias{add_unique_lineup_constraint}
\title{Unique Lineup Constraint}
\usage{
add_unique_lineup_constraint(model, roster_indx)
}
\arguments{
\item{model}{The model to further constrain}

\item{roster_indx}{the index of players to constrain}
}
\description{
Unique Lineup Constraint
}
\keyword{internal}
