% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-constraint.R
\name{apply_constraint,constraintClass-method}
\alias{apply_constraint,constraintClass-method}
\title{Function to apply constraints to a model}
\usage{
\S4method{apply_constraint}{constraintClass}(object, model, ...)
}
\arguments{
\item{object}{The constraint object}

\item{model}{Model object}

\item{...}{Additional arguments if required}
}
\value{
milp_model object
}
\description{
Function to apply constraints to a model
}
\keyword{internal}
