% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-player.R
\name{apply_variance,player_object-method}
\alias{apply_variance,player_object-method}
\title{apply variance value to player object}
\usage{
\S4method{apply_variance}{player_object}(object)
}
\arguments{
\item{object}{a player_object}
}
\description{
apply variance value to player object
}
