% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-lineups.R
\name{as.list.lineupClass}
\alias{as.list.lineupClass}
\title{Simple way to extract the lineup list to do whatever you want with it.}
\usage{
\method{as.list}{lineupClass}(object)
}
\description{
Simple way to extract the lineup list to do whatever you want with it.
}
