% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-model.R
\name{block_one_lineup}
\alias{block_one_lineup}
\title{Unique Lineup Constraint}
\usage{
block_one_lineup(model, roster_indx)
}
\arguments{
\item{model}{The model to further constrain}

\item{roster_indx}{the index of players to constrain}
}
\description{
Unique Lineup Constraint
}
\keyword{internal}
