% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-model.R
\name{build_base_model}
\alias{build_base_model}
\title{Build Base Model}
\usage{
build_base_model(size, pts)
}
\arguments{
\item{size}{number of units}

\item{pts}{vector of length 'size' containing points to use in objective function}
}
\description{
Build Base Model
}
