% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer-build-methods.R
\name{build_lineups,optimizer-method}
\alias{build_lineups,optimizer-method}
\title{Function to Generate lineups}
\usage{
\S4method{build_lineups}{optimizer}(object, num_lineups = 1, solver = "glpk", maximize = TRUE)
}
\arguments{
\item{object}{an S4 object of class Optimizer}

\item{num_lineups}{Number of lineups to generate}

\item{solver}{The solver method (defaults to 'glpk').}

\item{maximize}{Whether the model is intended to maximize (the default) or minimize the objective function}
}
\description{
Function to Generate lineups
}
