% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-model.R
\name{build_singlegame_model}
\alias{build_singlegame_model}
\title{Build Single Game / Showdown / Captain model}
\usage{
build_singlegame_model(
  size,
  team_vector,
  pts,
  mlt_mode = TRUE,
  maximize = TRUE
)
}
\arguments{
\item{size}{number of players}

\item{team_vector}{vector of teams}

\item{pts}{vector of length 'size' containing points to use in objective function}

\item{mlt_mode}{Logical. Determines whether captain mode is used.}

\item{maximize}{Whether to maximize the objective (if FALSE, the objective is minimized)}
}
\description{
Build Single Game / Showdown / Captain model
}
\keyword{internal}
