% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup-utils.R
\name{calc_jaccard_distance}
\alias{calc_jaccard_distance}
\title{Calculate a 'similarity measure'}
\usage{
calc_jaccard_distance(lineups)
}
\arguments{
\item{lineups}{a list of lineups}
}
\description{
Calculate a 'similarity measure'
}
\keyword{internal}
