% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lineups.R
\name{compare_lineups_with_actuals,lineupClass-method}
\alias{compare_lineups_with_actuals,lineupClass-method}
\alias{compare_lineups_with_actuals}
\title{Compare Lineups with Actuals}
\usage{
\S4method{compare_lineups_with_actuals}{lineupClass}(object, actuals, show_warnings = TRUE)
}
\arguments{
\item{object}{An object of class Lineups}

\item{actuals}{A data.frame (or data.table, or tibble) containing players and actual fantasy points.  See details.}

\item{show_warnings}{Flag indicating whether warnings should be shown (e.g., "Player X was not found in \code{actuals}").}
}
\value{
Data.table of lineup comparisons
}
\description{
This function can be used to determine what a lineup would have scored, given projections and actual
fantasy points. This makes back-testing different lineup parameters fast and easy.
}
\details{
The \code{actuals} data.frame needs to contain a column called \code{actuals} that contains actual fantasy points,
and at least one of \code{id} (player IDs associated with the model run), or \code{fullname}.  If both identifier columns
are present, \code{id} will be preferentially used.
}
