% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constr_force_opposing}
\alias{constr_force_opposing}
\title{Force Opposing Position constraint}
\usage{
constr_force_opposing(model, players, pos1, pos2)
}
\arguments{
\item{model}{Model object}

\item{players}{List of player objects}

\item{pos1}{Position for team}

\item{pos2}{Position for opposing team}
}
\description{
Force Opposing Position constraint
}
\details{
Accepts a single value for pos1 and pos2
}
\keyword{internal}
