% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constr_players_per_team}
\alias{constr_players_per_team}
\title{Players per Team Constraint}
\usage{
constr_players_per_team(model, players, players_per_team, exact = FALSE)
}
\arguments{
\item{model}{Model object}

\item{players}{List of player objects}

\item{players_per_team}{Positions to limit by team}

\item{exact}{Logical. Whether the team filter values should be honored exactly. Defaults to FALSE.}
}
\description{
Players per Team Constraint
}
\details{
Accepts either a named list of teams with numbers.
}
\keyword{internal}
