% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constr_restrict_opposing_position}
\alias{constr_restrict_opposing_position}
\title{Opposite Positions Constraint}
\usage{
constr_restrict_opposing_position(model, players, pos1, pos2)
}
\arguments{
\item{model}{Model object}

\item{players}{List of player objects}

\item{pos1}{Positions for set one}

\item{pos2}{Positions for set two}
}
\description{
Opposite Positions Constraint
}
\details{
The parameters \code{pos1} and \code{pos2} refer to the two sets of positions to keep
unmatched.
}
\keyword{internal}
