% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-constraint.R
\docType{class}
\name{constraintClass-class}
\alias{constraintClass-class}
\alias{.constraintClass}
\title{Constraint Class}
\description{
Constraint Class
}
\section{Slots}{

\describe{
\item{\code{constraint_name}}{A name for the constraint (used internally)}

\item{\code{fnc}}{A function to apply to the model}

\item{\code{args}}{A list containing arguments to pass to the function. Do not include a model class here, that is handled by the method \code{apply_constraint}.}
}}

\keyword{internal}
