% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer-build-methods.R
\name{construct_model,optimizer-method}
\alias{construct_model,optimizer-method}
\alias{construct_model}
\title{Method for constructing the optimization model}
\usage{
\S4method{construct_model}{optimizer}(object, maximize = TRUE)
}
\arguments{
\item{object}{an S4 object of class Optimizer}

\item{maximize}{Whether to maximize or minimize the objective function}
}
\description{
Method for constructing the optimization model
}
