% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lineups.R
\name{convert_lineup_to_vector,lineupClass-method}
\alias{convert_lineup_to_vector,lineupClass-method}
\alias{convert_lineup_to_vector}
\title{Convert lineup object to indexes}
\usage{
\S4method{convert_lineup_to_vector}{lineupClass}(x, opt)
}
\arguments{
\item{x}{lineupClass object}

\item{opt}{an optimizer model object}
}
\description{
Convert lineup object to indexes
}
\keyword{internal}
