% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-optimizer.R
\name{create_optimizer}
\alias{create_optimizer}
\title{Create an object of Optimizer}
\usage{
create_optimizer(
  site,
  sport,
  contest_type = "CLASSIC",
  players = list(),
  filepath = NULL
)
}
\arguments{
\item{site}{The site being used for optimization}

\item{sport}{The sport being optimized}

\item{contest_type}{The type of contest; determines base constraints (e.g., Classic, Showdown(DK)/Single(FD or Yahoo)). Default: CLASSIC}

\item{players}{List of players to build lineups from (defaults to empty list)}

\item{filepath}{Alternate method for adding players at time of creation. Passing a filepath will result in an internal call to
\code{add_players_from_csv}. If \code{players} is provided, filepath will be ignored.}
}
\description{
Create an object of Optimizer
}
\details{
This function is used to instantiate a new object of class \code{optimizer}, which is the
    central component of the dfsOptimizer package.
}
\examples{
mod <- create_optimizer(site = 'DRAFTKINGS', sport = 'HOCKEY')

}
