% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{.add_opposing_position_constraint}
\alias{.add_opposing_position_constraint}
\title{Opposite Positions Constraint}
\usage{
.add_opposing_position_constraint(model, pos1, pos2, players)
}
\arguments{
\item{model}{Model object}

\item{pos1}{Positions for set one}

\item{pos2}{Positions for set two}

\item{players}{List of Player objects}
}
\description{
Opposite Positions Constraint
}
\details{
The parameters \code{pos1} and \code{pos2} refer to the two sets of positions to keep
unmatched.
}
\keyword{internal}
