% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{.add_same_team_stack}
\alias{.add_same_team_stack}
\title{Adds Same-Team stacks}
\usage{
.add_same_team_stack(model, positions, players)
}
\arguments{
\item{model}{Model object}

\item{positions}{Positions for that should be stacked within a single team}

\item{players}{List of Player objects}
}
\description{
Adds Same-Team stacks
}
\details{

}
\keyword{(probably}
\keyword{PER}
\keyword{add}
\keyword{and}
\keyword{constraints}
\keyword{internal}
\keyword{need}
\keyword{number}
\keyword{of}
\keyword{one}
\keyword{players}
\keyword{positions,}
\keyword{set}
\keyword{team.)}
\keyword{teams}
\keyword{the}
\keyword{to}
