% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lineups.R
\name{export_lineups,lineupClass-method}
\alias{export_lineups,lineupClass-method}
\title{Export Lineups to CSV}
\usage{
\S4method{export_lineups}{lineupClass}(object, file = "lineup_export.csv")
}
\arguments{
\item{object}{An object of class Lineups}

\item{file}{A character string indicating the file to write to}
}
\description{
Export Lineups to CSV
}
