% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-optimizer.R
\name{extract_player_fpts,optimizer-method}
\alias{extract_player_fpts,optimizer-method}
\title{Extracting Fantasy Points from Player objects}
\usage{
\S4method{extract_player_fpts}{optimizer}(object)
}
\arguments{
\item{object}{an optimizer object}
}
\description{
Extracting Fantasy Points from Player objects
}
