% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{extract_player_fpts}
\alias{extract_player_fpts}
\title{Extracting Fantasy Points from Player objects}
\usage{
extract_player_fpts(object)
}
\arguments{
\item{object}{an optimizer object}
}
\description{
Extracting Fantasy Points from Player objects
}
