% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{force_opposing_positions,optimizer-method}
\alias{force_opposing_positions,optimizer-method}
\title{Force Opposing Positions}
\usage{
\S4method{force_opposing_positions}{optimizer}(object, pos1, pos2)
}
\arguments{
\item{object}{An optimizer object}

\item{pos1}{Position for team one}

\item{pos2}{Position for opposing team}
}
\value{
Updated optimizer object
}
\description{
Force Opposing Positions
}
\details{
Forces positions from opposing teams from being included in lineups. Currently accepts a maximum
of two positions.
}
\examples{
\dontrun{
opt <- create_optimizer(site = 'DRAFTKINGS', sport = 'FOOTBALL', contest_type = 'CLASSIC')
opt <- add_players_from_csv(object = opt, filepath = '/Path/to/file.csv')

# Force lineup to include a QB and a WR from the opposing team
opt <- force_opposing_positions(object = opt, pos1 = 'QB', pos2 = 'WR')
}

}
