% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lineups.R
\name{format_for_upload,lineupClass-method}
\alias{format_for_upload,lineupClass-method}
\title{Function for getting lineups into uploadable format}
\usage{
\S4method{format_for_upload}{lineupClass}(object)
}
\arguments{
\item{object}{An object of class Lineups}
}
\description{
Function for getting lineups into uploadable format
}
