% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-game-info.R
\docType{class}
\name{gameInfo-class}
\alias{gameInfo-class}
\alias{gameInfo}
\title{S4 Class: Game Info}
\description{
Information about an individual player's game
}
\section{Slots}{

\describe{
\item{\code{home_team}}{The home team}

\item{\code{away_team}}{the away team}

\item{\code{shortname}}{Short name for the game (see Details).}

\item{\code{start_time}}{Start time for the game}

\item{\code{is_started}}{a flag determining whether the game has started or not}
}}

