% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-config.R
\name{generate_base_model}
\alias{generate_base_model}
\title{Generate a basic model from a construction object}
\usage{
generate_base_model(object)
}
\description{
Generate a basic model from a construction object
}
