% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup-utils.R
\name{get_exposure_summary}
\alias{get_exposure_summary}
\title{Calculate Exposure percentages across lineups}
\usage{
get_exposure_summary(lineups)
}
\arguments{
\item{lineups}{a list of lineups}
}
\description{
Calculate Exposure percentages across lineups
}
\keyword{internal}
