% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{get_player_data,optimizer-method}
\alias{get_player_data,optimizer-method}
\alias{get_player_data}
\title{Get Player data from optimizer object}
\usage{
\S4method{get_player_data}{optimizer}(object)
}
\arguments{
\item{object}{An optimizer object}
}
\value{
Data.table containing information about the players.
}
\description{
Get Player data from optimizer object
}
\details{
In addition to the optimizer object, this method can be run on a player object
to provide details about that single player.
}
