% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-player.R
\name{get_player_data,player_object-method}
\alias{get_player_data,player_object-method}
\title{player_object formatter method}
\usage{
\S4method{get_player_data}{player_object}(object)
}
\arguments{
\item{object}{Object of class \code{player_object}}
}
\description{
player_object formatter method
}
