% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{get_player_id,optimizer-method}
\alias{get_player_id,optimizer-method}
\title{Get a Player Id}
\usage{
\S4method{get_player_id}{optimizer}(object, name, team, position)
}
\arguments{
\item{object}{Object of class Optimizer}

\item{name}{Full name of player}

\item{team}{team abbreviation of player (Not required)}

\item{position}{position of player (Not required)}
}
\description{
Get a Player Id
}
\details{
\code{team} and \code{position} can be included to differentiate between two players with the same name, but who play for different teams and/or at different positions.
}
