% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-optimizer.R
\name{get_player_index,optimizer-method}
\alias{get_player_index,optimizer-method}
\title{Method for getting player index (Deprecated?)}
\usage{
\S4method{get_player_index}{optimizer}(object)
}
\arguments{
\item{object}{an S4 object of class Optimizer}
}
\description{
Method for getting player index (Deprecated?)
}
