% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup-utils.R
\name{get_player_summary}
\alias{get_player_summary}
\title{Calculate Player percentages across lineups}
\usage{
get_player_summary(lineups)
}
\arguments{
\item{lineups}{a list of lineups}
}
\description{
Calculate Player percentages across lineups
}
\keyword{internal}
