% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-players-from-csv.R
\name{get_players_from_csv,YahooOptim-method}
\alias{get_players_from_csv,YahooOptim-method}
\title{Internal Method for Subclasses}
\usage{
\S4method{get_players_from_csv}{YahooOptim}(object, path)
}
\arguments{
\item{object}{Optimizer subclass YahooOptim}

\item{path}{Path to CSV}
}
\description{
Internal Method for Subclasses
}
