% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-players-from-csv.R
\name{get_players_from_csv,YahooOptim-method}
\alias{get_players_from_csv,YahooOptim-method}
\alias{get_players_from_csv,FanduelOptim-method}
\title{Create Players from CSV}
\usage{
\S4method{get_players_from_csv}{YahooOptim}(object, path)

\S4method{get_players_from_csv}{FanduelOptim}(object, path)
}
\arguments{
\item{object}{Optimizer object}

\item{path}{Path to CSV or text file}
}
\description{
Create Players from CSV

Create Players from CSV
}
\details{
Reads a CSV from a specified site, and generates objects of Player class (one per row)

Reads a CSV from a specified site, and generates objects of Player class (one per row)
}
