% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-players-from-df.R
\name{get_players_from_data_frame}
\alias{get_players_from_data_frame}
\title{Get Players from data.frame}
\usage{
get_players_from_data_frame(df)
}
\arguments{
\item{df}{A data.frame with columns for player related information.  This can contain any/all of the player related info.
If a column coincides with a slot in objects of class \code{player_object}, it will be included when the
data.frame is converted into unique player objects. The one exception is slot \code{game_info}, which is
ignored in the conversion.}
}
\value{
A list containing values of class \code{player_objects}, one per player (i.e., row) in the data set.
}
\description{
Get Players from data.frame
}
\details{
The following columns that are required, in order to avoid issues with the optimizer model:
\itemize{
  \item{first_name}
  \item{last_name}
  \item{team}
  \item{position}
  \item{salary}
  \item{fpts}
}
You'll get an error if these aren't included.  If you want to know what else you can include, run \code{list_player_attributes()}
}
