% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup-utils.R
\name{get_team_summary}
\alias{get_team_summary}
\title{Calculate Team percentages across lineups}
\usage{
get_team_summary(lineups)
}
\arguments{
\item{lineups}{a list of lineups}
}
\description{
Calculate Team percentages across lineups
}
\keyword{internal}
