% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{include_constraint}
\alias{include_constraint}
\title{Including Constraints}
\usage{
include_constraint(x, constraint_object)
}
\arguments{
\item{x}{Optimizer object}

\item{constraint_object}{an object of constraint class}
}
\description{
Including Constraints
}
\keyword{internal}
