% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-lineups.R
\docType{class}
\name{lineupClass-class}
\alias{lineupClass-class}
\title{S4 Lineup class}
\description{
S4 Lineup class
}
\section{Slots}{

\describe{
\item{\code{num_lineups}}{Number of lineups}

\item{\code{lineups}}{A list of lineups}
}}

