% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player-utils.R
\name{link_players_on_same_team}
\alias{link_players_on_same_team}
\title{Get all combinations of players on same team}
\usage{
link_players_on_same_team(players, positions)
}
\arguments{
\item{players}{a list of Player objects}

\item{positions}{a character vector of positions to align}
}
\description{
Get all combinations of players on same team
}
\keyword{internal}
