% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-player.R
\name{list_player_attributes}
\alias{list_player_attributes}
\title{List player attributes}
\usage{
list_player_attributes()
}
\value{
A vector of the attributes available for player objects (i.e., the values the user can set for an individual player).
}
\description{
List player attributes
}
