% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{lock_players_by_id,optimizer-method}
\alias{lock_players_by_id,optimizer-method}
\title{Function to block players by ID}
\usage{
\S4method{lock_players_by_id}{optimizer}(object, player_ids)
}
\arguments{
\item{object}{an S4 object of class Optimizer}

\item{player_ids}{IDs of players to block}
}
\value{
Updated optimizer object
}
\description{
Function to block players by ID
}
