% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhl_players}
\alias{nhl_players}
\title{A set of NHL Player Data to use for testing}
\format{
A data table with 308 rows and 8 variables:
\describe{
  \item{id}{ID of player}
  \item{first_name}{Player's first name}
  \item{last_name}{Player's last name}
  \item{team}{team}
  \item{position}{Roster position}
  \item{salary}{How much the player costs}
  \item{fpts}{Projected fantasy points}
  \item{game_info}{Info about the player's game, such as home/away teams and time}
}
}
\usage{
nhl_players
}
\description{
A dataset containing the names, ids, teams, and other attributes of
308 NHL Players for a given Draftkings Contest.
}
\keyword{datasets}
