% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-optim-model.R
\docType{class}
\name{optim_model-class}
\alias{optim_model-class}
\alias{optim_model}
\title{S4 Class of Optimization Model}
\description{
S4 Class of Optimization Model
}
\section{Slots}{

\describe{
\item{\code{mod}}{The model object}

\item{\code{solver}}{Method for solving; currently only 'glpk' is available}

\item{\code{flex_positions}}{Character array of flex positions}
}}

