% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-optimizer.R
\name{optimize,optimizer-method}
\alias{optimize,optimizer-method}
\title{Function to Generate lineups}
\usage{
\S4method{optimize}{optimizer}(object, num_lineups = 1)
}
\arguments{
\item{object}{an S4 object of class Optimizer}

\item{num_lineups}{Number of lineups to generate}
}
\description{
Function to Generate lineups
}
