% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-optimizer.R
\docType{class}
\name{optimizer-class}
\alias{optimizer-class}
\alias{.optimizer}
\title{S4 Class For Optimizer}
\description{
S4 Class For Optimizer
}
\section{Slots}{

\describe{
\item{\code{site}}{The site being used for optimization}

\item{\code{sport}}{The sport being optimized}

\item{\code{contest_type}}{The type of contest; determines base constraints (e.g., Classic, Showdown/Single-Game)}

\item{\code{players}}{List of players to build lineups from}

\item{\code{model}}{The optimization model}

\item{\code{maximize}}{Logical, whether to maximize or minimize the objective function}
}}

