% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-players-from-csv.R
\name{parse_game_info}
\alias{parse_game_info}
\title{Function to parse string for game info (Internal)}
\usage{
parse_game_info(str)
}
\arguments{
\item{str}{A String containing game info}
}
\description{
Function to parse string for game info (Internal)
}
\details{
Used internally to aid player creation
}
\keyword{internal}
