% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup-utils.R
\name{parse_roster_key}
\alias{parse_roster_key}
\title{Parse roster key item}
\usage{
parse_roster_key(rk)
}
\arguments{
\item{rk}{Roster key from configuration object}
}
\description{
Parse roster key item
}
\details{
This function will take all single and multiplayer slots from the roster key, and
    build a table containing mins/maxes for each position.  These are then used to set the
    value of how many of each position is allowable in a lineup.
}
\keyword{internal}
