% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-player.R
\name{player}
\alias{player}
\title{Constructor Function for Player class}
\usage{
player(
  id,
  first_name,
  last_name,
  team,
  position,
  depth = 1L,
  salary,
  fpts,
  locked = FALSE,
  blocked = FALSE,
  is_injured = FALSE,
  game_info
)
}
\arguments{
\item{id}{player ID}

\item{first_name}{first name}

\item{last_name}{last name}

\item{team}{team (Required)}

\item{position}{position (Required)}

\item{depth}{Position on depth chart or roster order (like batting order) (Default 1)}

\item{salary}{salary}

\item{fpts}{fantasy points}

\item{locked}{Whether the player should be locked to all lineups (Default FALSE)}

\item{blocked}{Whether to omit player from all lineups (Default FALSE)}

\item{is_injured}{injury flag (Default FALSE)}

\item{game_info}{information about the game (Default is an empty game_info object)}
}
\description{
Constructor Function for Player class
}
\details{
Function that initializes a Player object.
}
