% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-player.R
\docType{class}
\name{player_object-class}
\alias{player_object-class}
\title{S4 Player Class}
\description{
S4 Player Class
}
\section{Slots}{

\describe{
\item{\code{id}}{player ID}

\item{\code{first_name}}{first name}

\item{\code{last_name}}{last name}

\item{\code{fullname}}{full name}

\item{\code{team}}{team}

\item{\code{position}}{position}

\item{\code{depth}}{Position on depth chart or roster order (like batting order)}

\item{\code{salary}}{salary}

\item{\code{fpts}}{fantasy points}

\item{\code{locked}}{Whether the player should be locked to all lineups}

\item{\code{blocked}}{whether to omit player from all lineups}

\item{\code{is_injured}}{injury flag}

\item{\code{min_exposure}}{Single player minimum exposure}

\item{\code{max_exposure}}{Single player maximum exposure}

\item{\code{variance}}{Single player variance}

\item{\code{game_info}}{information about the game}
}}

\keyword{internal}
