% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{remove_player,optimizer-method}
\alias{remove_player,optimizer-method}
\title{Method for removing a player from Optimizer object}
\usage{
\S4method{remove_player}{optimizer}(object, id)
}
\arguments{
\item{object}{an S4 optimizer object}

\item{id}{ID of player to remove}
}
\value{
Updated optimizer object
}
\description{
Method for removing a player from Optimizer object
}
