% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup-utils.R
\name{reorder_lineup}
\alias{reorder_lineup}
\title{Reorder lineups}
\usage{
reorder_lineup(lineup, new_order)
}
\arguments{
\item{lineup}{the lineup output}

\item{new_order}{The desired ordered output of the table (usually a character vector of positions equal to the nrow(lineup))}
}
\description{
Reorder lineups
}
\keyword{internal}
