% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{restrict_opposing_positions,optimizer-method}
\alias{restrict_opposing_positions,optimizer-method}
\title{Restrict Opposing Positions}
\usage{
\S4method{restrict_opposing_positions}{optimizer}(object, pos1, pos2)
}
\arguments{
\item{object}{An optimizer object}

\item{pos1}{Positions for set one}

\item{pos2}{Positions for set two}
}
\value{
Updated optimizer object
}
\description{
Restrict Opposing Positions
}
\details{
Prevents positions from opposing teams from being included in lineups.
}
\examples{
\dontrun{
opt <- create_optimizer(site = 'DRAFTKINGS', sport = 'HOCKEY', contest_type = 'CLASSIC')
opt <- add_players_from_csv(object = opt, filepath = '/Path/to/file.csv')

# Restrict lineup from having Skaters and Goalies from opposite teams
opt <- restrict_opposing_positions(object = opt, pos1 = c('C','W','D'), pos2 = 'G')
}

}
