% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-player.R
\name{salary,player_object-method}
\alias{salary,player_object-method}
\title{Accessing Data from Player Objects}
\usage{
\S4method{salary}{player_object}(x)
}
\arguments{
\item{x}{player_object}
}
\description{
Accessing Data from Player Objects
}
