% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-player.R
\name{set_as_injured}
\alias{set_as_injured}
\title{Internal function for setting player as injured}
\usage{
set_as_injured(object)
}
\description{
Internal function for setting player as injured
}
\keyword{internal}
