% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{set_flex_positions,optimizer-method}
\alias{set_flex_positions,optimizer-method}
\title{Set the FLEX/UTIL position list}
\usage{
\S4method{set_flex_positions}{optimizer}(object, positions)
}
\arguments{
\item{object}{An optimizer object}

\item{positions}{Value to set the roster size (subject to validity checks).}
}
\value{
Updated Optimizer object
}
\description{
Method for setting FLEX/UTIL positions. Can be used to limit FLEX/UTIL to a user-specified
   set of possible positions (see examples).
}
\examples{
\dontrun{
opt <- optimizer(site = 'DRAFTKINGs', sport = 'HOCKEY')
opt <- add_players_from_csv(object = opt, filepath = '/Path/to/file.csv')

# Default FLEX is c('C','W','D') -- here we'll remove D
opt <- set_flex_positions(object = opt, positions = c('C','W'))
}

}
