% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-player.R
\name{set_fpts<-,player_object-method}
\alias{set_fpts<-,player_object-method}
\alias{set_fpts}
\title{Setting Fpts}
\usage{
\S4method{set_fpts}{player_object}(object) <- value
}
\description{
Setting Fpts
}
