% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-player.R
\name{set_fpts<-}
\alias{set_fpts<-}
\title{Setting Fpts}
\usage{
set_fpts(object) <- value
}
\arguments{
\item{object}{A player_object}

\item{value}{Value of FPTs}
}
\description{
Setting Fpts
}
\keyword{internal}
