% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer-player-setting.R
\name{set_fpts_by_id,optimizer-method}
\alias{set_fpts_by_id,optimizer-method}
\alias{set_fpts_by_id}
\title{Method for updating fantasy points in an object}
\usage{
\S4method{set_fpts_by_id}{optimizer}(object, id, fpts)
}
\arguments{
\item{object}{An object of class Optimizer}

\item{id}{A Player ID to update}

\item{fpts}{Value for slot \code{fpts} of Player object}
}
\value{
Updated optimizer object
}
\description{
Method for updating fantasy points in an object
}
