% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{set_global_variance,optimizer-method}
\alias{set_global_variance,optimizer-method}
\alias{set_global_variance}
\title{Set the Global Variance}
\usage{
\S4method{set_global_variance}{optimizer}(object, variance)
}
\arguments{
\item{object}{An optimizer object}

\item{variance}{Value to set the variance}
}
\value{
Updated Optimizer object
}
\description{
Method to set the global variance of an optimizer model.
}
\examples{
\dontrun{
# Set Optimizer global variance to 10\%
opt <- opt \%>\%
   set_global_variance(variance = .10)
}

}
