% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{set_max_budget,optimizer-method}
\alias{set_max_budget,optimizer-method}
\alias{set_max_budget}
\title{Set the maximium budget for the model}
\usage{
\S4method{set_max_budget}{optimizer}(object, max_budget)
}
\arguments{
\item{object}{An optimizer object}

\item{max_budget}{The max budget value}
}
\value{
Updated optimizer object
}
\description{
Used to set a upper-bound budget for all lineups
}
\examples{
\dontrun{
# Set the maximum budget to 45000
opt <- opt \%>\%
  set_max_budget(45000)
}

}
