% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{set_max_exposure,optimizer-method}
\alias{set_max_exposure,optimizer-method}
\title{Set the Global Max Exposure}
\usage{
\S4method{set_max_exposure}{optimizer}(object, exposure)
}
\arguments{
\item{object}{An optimizer object}

\item{exposure}{Value to set the roster size (subject to validity checks).}
}
\value{
Updated Optimizer object
}
\description{
Method to set the global max exposure of an optimizer model.
}
