% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{set_max_overlap,optimizer-method}
\alias{set_max_overlap,optimizer-method}
\title{Set the maximum overlap}
\usage{
\S4method{set_max_overlap}{optimizer}(object, overlap)
}
\arguments{
\item{object}{An optimizer object}

\item{overlap}{Maximum player overlap across lineups}
}
\value{
Updated Optimizer object
}
\description{
Method to set the max overlap of an optimizer model.
}
\examples{
\dontrun{
opt <- create_optimizer(site = 'DRAFTKINGs', sport = 'HOCKEY', contest_type = 'CLASSIC')
opt <- add_players_from_csv(object = opt, filepath = '/Path/to/file.csv')

# update the max overlap
opt <- set_max_overlap(object = opt, overlap = 6)
}

}
