% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{set_max_teams,optimizer-method}
\alias{set_max_teams,optimizer-method}
\alias{set_max_teams}
\title{Set minimum team requirement}
\usage{
\S4method{set_max_teams}{optimizer}(object, max_teams)
}
\arguments{
\item{object}{An optimizer object}

\item{max_teams}{The maximum number of teams to include}
}
\description{
Set minimum team requirement
}
\details{
WARNING: You can very easily make your lineup ineligible for a site by doing things
like \code{set_max_teams(object, max_teams = 1)}
}
\examples{
\dontrun{
# Set minimum number of teams to 6
opt <- set_max_teams(opt, max_teams = 6)
}

}
