% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{set_min_budget,optimizer-method}
\alias{set_min_budget,optimizer-method}
\title{Set the minimum budget for the model}
\usage{
\S4method{set_min_budget}{optimizer}(object, min_budget)
}
\arguments{
\item{object}{An optimizer object}

\item{min_budget}{The minimum budget value}
}
\value{
Updated optimizer object
}
\description{
Used to set a lower-bound budget for all lineups
}
\examples{
\dontrun{
# Set the minimum budget to 40000
opt <- opt \%>\%
  set_min_budget(40000)
}

}
