% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{set_min_teams,optimizer-method}
\alias{set_min_teams,optimizer-method}
\alias{set_min_teams}
\title{Set minimum team requirement}
\usage{
\S4method{set_min_teams}{optimizer}(object, min_teams)
}
\arguments{
\item{object}{An optimizer object}

\item{min_teams}{The minimum number of teams to include}
}
\description{
Set minimum team requirement
}
\details{
WARNING: You can very easily make your lineup ineligible for a site by doing things
like \code{set_min_teams(object, min_teams = 1)}
}
\examples{
\dontrun{
# Set minimum number of teams to 6
opt <- set_min_teams(opt, min_teams = 6)
}

}
