% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{set_multiplier_position,SingleGameOptim-method}
\alias{set_multiplier_position,SingleGameOptim-method}
\alias{set_multiplier_position}
\title{Set Possible positions for Multiplier slot}
\usage{
\S4method{set_multiplier_position}{SingleGameOptim}(object, positions)
}
\arguments{
\item{object}{An optimizer object}

\item{positions}{Value to set the roster size (subject to validity checks).}
}
\value{
Updated Optimizer object
}
\description{
Method for setting positions for Single Game multiplier positions (e.g., CPT in Draftkings Captain-mode).
   Can be used to limit the multiplier position to a user-specified set of possible positions (see examples).
}
\examples{
\dontrun{
opt <- create_optimizer(site = 'DRAFTKINGS', sport = 'HOCKEY', contest_type = 'SHOWDOWN')
opt <- add_players_from_csv(object = opt, filepath = '/Path/to/file.csv')

# Default Multiplier-position options are c('C','W','D','G')
# Remove D and G
opt <- set_multiplier_position(object = opt, positions = c('C','W'))
}

}
