% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{set_player_max_exp,optimizer-method}
\alias{set_player_max_exp,optimizer-method}
\alias{set_player_max_exp}
\title{Set a player's max exposure by ID}
\usage{
\S4method{set_player_max_exp}{optimizer}(object, id, exp)
}
\arguments{
\item{object}{An optimizer object}

\item{id}{The ID of a player}

\item{exp}{A value of exposure}
}
\value{
Updated Optimizer object
}
\description{
Set a player's max exposure by ID
}
\examples{
\dontrun{
# Set Carey Price to have a maximum exposure of 75\%
ID <- get_player_id(opt, 'Carey Price')
opt <- opt \%>\% set_player_max_exp(id = ID, exp = .75)
}

}
