% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-player.R
\name{set_player_max_exposure}
\alias{set_player_max_exposure}
\title{Player_object max exposure}
\usage{
set_player_max_exposure(object, exposure)
}
\arguments{
\item{object}{A Player_object}

\item{exposure}{Value of exposure}
}
\description{
Player_object max exposure
}
\keyword{internal}
