% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer-player-setting.R
\name{set_player_min_exp,optimizer-method}
\alias{set_player_min_exp,optimizer-method}
\alias{set_player_min_exp}
\title{Set a player's Minimum exposure by ID}
\usage{
\S4method{set_player_min_exp}{optimizer}(object, id, exp)
}
\arguments{
\item{object}{An optimizer object}

\item{id}{The ID of a player}

\item{exp}{A value of exposure}
}
\value{
Updated Optimizer object
}
\description{
Set a player's Minimum exposure by ID
}
\examples{
\dontrun{
# Set Patrick Mahomes to have a minimum exposure of 75\%
ID <- get_player_id(opt, 'Patrick Mahomes')
opt <- opt \%>\% set_player_min_exp(id = ID, exp = .75)
}

}
