% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{set_player_variance,optimizer-method}
\alias{set_player_variance,optimizer-method}
\alias{set_player_variance}
\title{Set a player's max exposure by ID}
\usage{
\S4method{set_player_variance}{optimizer}(object, id, variance)
}
\arguments{
\item{object}{An optimizer object}

\item{id}{The ID of a player}

\item{variance}{A value of variance}
}
\value{
Updated Optimizer object
}
\description{
Set a player's max exposure by ID
}
\examples{
\dontrun{
# Set Carey Price's fpts to vary by plus-or-minus 25\%
ID <- get_player_id(opt, 'Carey Price')
opt <- opt \%>\% set_player_variance(id = ID, variance = .25)
}

}
