% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{set_roster_size,optimizer-method}
\alias{set_roster_size,optimizer-method}
\title{Set the Roster Size}
\usage{
\S4method{set_roster_size}{optimizer}(object, value)
}
\arguments{
\item{object}{An optimizer object}

\item{value}{Value to set the roster size (subject to validity checks).}
}
\value{
Updated Optimizer object
}
\description{
Method to set the roster size requirement of an optimizer model.
}
