% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-player.R
\name{set_variance}
\alias{set_variance}
\title{Setting Variance}
\usage{
set_variance(object, variance)
}
\arguments{
\item{object}{A Player_object}

\item{variance}{Value of variance}
}
\description{
Setting Variance
}
