% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-lineups.R
\name{summary.lineupClass}
\alias{summary.lineupClass}
\title{Summary for lineupClass}
\usage{
\method{summary}{lineupClass}(object, ...)
}
\arguments{
\item{object}{lineupClass object}

\item{...}{Unused. Included for S3 consistency}
}
\description{
Summary for lineupClass
}
