% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{toggle_multiplier_mode}
\alias{toggle_multiplier_mode}
\title{Toggle Multiplier Mode}
\usage{
toggle_multiplier_mode(object)
}
\arguments{
\item{object}{An object of class Optimizer}
}
\description{
Toggle Multiplier Mode
}
\details{
Toggles the Multiplier mode (where a player's salary and fpts are multiplied by a given value, usually 1.5)
    for single-game / showdown contest types.  For Classic contest_types, this function has no effect.
}
