% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{update_fpts,optimizer-method}
\alias{update_fpts,optimizer-method}
\title{Method for updating fantasy points in an object}
\usage{
\S4method{update_fpts}{optimizer}(object, fpts_data)
}
\arguments{
\item{object}{An object of class Optimizer}

\item{fpts_data}{a data.frame containing players and points. See details.}
}
\value{
Updated optimizer object
}
\description{
Method for updating fantasy points in an object
}
\details{
The data.frame passed in fpts_data should contain two columns - \code{id} and \code{fpts}.
}
