% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{add_block_constraint}
\alias{add_block_constraint}
\title{Block Players Constraint}
\usage{
add_block_constraint(model, block_vector)
}
\arguments{
\item{model}{Model object}

\item{block_vector}{Vector where 1 indicates 'block'}
}
\description{
Block Players Constraint
}
\keyword{internal}
