% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-optimizer.R
\name{add_player,optimizer-method}
\alias{add_player,optimizer-method}
\title{Method for adding a player to Optimizer object}
\usage{
\S4method{add_player}{optimizer}(object, pl)
}
\arguments{
\item{object}{an S4 optimizer object}

\item{pl}{an object of class Player}
}
\description{
Method for adding a player to Optimizer object
}
