% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-optimizer.R
\name{add_players_from_csv,optimizer-method}
\alias{add_players_from_csv,optimizer-method}
\alias{add_players_from_csv}
\title{Add players to optimizer from CSV}
\usage{
\S4method{add_players_from_csv}{optimizer}(object, filepath, custom = FALSE)
}
\arguments{
\item{object}{An optimizer model object}

\item{filepath}{Filepath location of the CSV}

\item{custom}{Set to TRUE to use user-generated CSVs with player data; the data will be
parsed using \code{get_players_from_date_frame}}
}
\value{
Optimizer model object with slot \code{players} filled
}
\description{
Add players to optimizer from CSV
}
\examples{
\dontrun{
opt <- create_optimizer(site = 'DRAFTKINGS', sport = 'HOCKEY', contest_type = 'CLASSIC')
opt <- add_players_from_csv(object = opt, filepath = '/Path/to/file.csv')
}

}
