% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-lineups.R
\name{as.list.lineupClass}
\alias{as.list.lineupClass}
\title{extract lineup list}
\usage{
\method{as.list}{lineupClass}(x, ...)
}
\arguments{
\item{x}{lineupClass object}

\item{...}{Unused. Included for S3 consistency}
}
\description{
extract lineup list
}
