% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer-build-methods.R
\name{build_lineups,optimizer-method}
\alias{build_lineups,optimizer-method}
\title{Function to Generate lineups}
\usage{
\S4method{build_lineups}{optimizer}(object, num_lineups = 1, solver = "glpk")
}
\arguments{
\item{object}{an S4 object of class Optimizer}

\item{num_lineups}{Number of lineups to generate}
}
\description{
Function to Generate lineups
}
