% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{.add_team_stack}
\alias{.add_team_stack}
\title{Adds Same-Team stacks}
\usage{
.add_team_stack(optObj, positions, nstacks = 1)
}
\arguments{
\item{optObj}{Optimizer object}

\item{positions}{Positions for that should be stacked within a single team}

\item{nstacks}{Number of stacks to try to include (Default is 1)}
}
\description{
Adds Same-Team stacks
}
\keyword{internal}
