% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{.add_team_stack}
\alias{.add_team_stack}
\title{Adds Same-Team stacks}
\usage{
.add_team_stack(model, positions)
}
\arguments{
\item{model}{Model object}

\item{positions}{Positions for that should be stacked within a single team}
}
\description{
Adds Same-Team stacks
}
\keyword{internal}
