% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_position_indicator}
\alias{make_position_indicator}
\title{make position indicator}
\usage{
make_position_indicator(posvec, target)
}
\arguments{
\item{posvec}{vector of positions}

\item{target}{position to flag}
}
\description{
make position indicator
}
