% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-config.R
\docType{class}
\name{optimConfig-class}
\alias{optimConfig-class}
\alias{.optimConfig}
\title{S4 Class optimConfig}
\description{
S4 Class optimConfig
}
\section{Slots}{

\describe{
\item{\code{budget}}{Lineup budget}

\item{\code{min_budget}}{the minimum budget (Default: 0)}

\item{\code{roster_size}}{Roster limitation (How many players allowed)}

\item{\code{min_team_req}}{Number of teams required to be represented in the lineup}

\item{\code{max_players_per_team}}{Maximum number of players from any one team}

\item{\code{roster_key}}{List containing roster positions and the number of each position required}

\item{\code{flex_position}}{named identifer for the flex positions}

\item{\code{max_exposure}}{Maximum exposure for individual players (global)}

\item{\code{variance}}{Percentage variance for fantasy points -- used to add randomness to the model.}

\item{\code{constraints}}{A \code{list} containing additional constraint objects}
}}

