% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup-utils.R
\name{reorder_lineup}
\alias{reorder_lineup}
\title{Reorder lineups}
\usage{
reorder_lineup(lineup, config)
}
\arguments{
\item{lineup}{the lineup output}

\item{config}{The config slot of the current optimizer. Used to determine the order, and eligible positions.}
}
\description{
Reorder lineups
}
\keyword{internal}
